us_market_dashboard/requirements.txt
streamlit==1.35.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
requests==2.32.3
feedparser==6.0.11
praw==7.7.1
fredapi==0.5.2
ta==0.11.0
scipy==1.13.0
openai==1.30.0
python-dotenv==1.0.1
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.2.2
stockstats==0.6.2
mplfinance==0.12.10b0
matplotlib==3.9.0
seaborn==0.13.2
streamlit-extras==0.4.2
streamlit-option-menu==0.3.12
altair==5.3.0
